+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## [1.0.1]
+- Added `SimpleCalculator.hideSurroundingBorder`.
+- Added `CalculatorThemeData.borderWidth`.
+- Fixed a state related issue.
+
 ## [1.0.0]
 - First Release.